refactor: extract warn applier into lib function
This commit will move the contents of the warn command into a new
function exported in lib/. Thus, it will be possible to use this
business logic in other parts that are not the warn command.
